requirements.txt
authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click-spinner==0.1.10
click==8.1.3
cloup==1.0.0
cryptography==37.0.4
flask-cors==3.0.10
flask==2.1.3
gevent==21.12.0
greenlet==1.1.2; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0; python_version < '3.10'
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.1
marshmallow-dataclass==8.5.8
marshmallow-enum==1.5.1
marshmallow==3.17.0
mypy-extensions==0.4.3
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pyjwt==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-http-client==3.3.7
pytz==2022.1
requests==2.28.1
sendgrid==6.9.7
six==1.16.0
sqlalchemy==1.4.39
starkbank-ecdsa==2.0.3
typeguard==2.13.3
typing-extensions==4.3.0
typing-inspect==0.7.1
urllib3==1.26.10
werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
